feat: added ranked mixt push and captains mix

+++ bot/src/interactions/commands/setup_mix.ts
@@ -3,6 +3,7 @@ import { ChatInputCommandInteraction } from "discord.js";
 import { BOT_ADMIN_ROLE, MAX_LINEUP_NAME_LENGTH } from "../../constants";
 import { ICommandHandler } from "../../handlers/commandHandler";
 import { LineupQueue } from "../../mongoSchema";
+import { authorizationService } from "../../services/authorizationService";
 import { interactionUtils } from "../../services/interactionUtils";
 import { matchmakingService } from "../../services/matchmakingService";
 import { LINEUP_TYPE_MIX, LINEUP_VISIBILITY_PUBLIC, LINEUP_VISIBILITY_TEAM, teamService } from "../../services/teamService";
@@ -38,8 +39,10 @@ export default {
                 .addChoices(
                     { name: 'Team', value: LINEUP_VISIBILITY_TEAM },
                     { name: 'Public', value: LINEUP_VISIBILITY_PUBLIC }
-                )
-            ),
+                ))
+            .addBooleanOption(option => option.setName('ranked')
+                .setRequired(false)
+                .setDescription('Indicates if this lineup is ranked and should update ratings')),
     authorizedRoles: [BOT_ADMIN_ROLE],
     async execute(interaction: ChatInputCommandInteraction) {
         let team = await teamService.findTeamByGuildId(interaction.guildId!)
@@ -57,13 +60,19 @@ export default {
             return
         }
 
+        const ranked = interaction.options.getBoolean("ranked") !== false
+        if (ranked && !authorizationService.isOfficialDiscord(interaction.guildId!)) {
+            await interaction.reply({ content: "⛔ Only official community discords can create ranked mixes", ephemeral: true })
+            return
+        }
+
         const lineupSize = interaction.options.getInteger("size")!
         const visibility = interaction.options.getString("visibility") || LINEUP_VISIBILITY_TEAM
-        const lineup = teamService.createLineup(interaction.channelId, lineupSize, lineupName, true, false, team, LINEUP_TYPE_MIX, visibility)
+        const lineup = teamService.createLineup(interaction.channelId, lineupSize, lineupName, true, ranked, team, LINEUP_TYPE_MIX, visibility)
         await teamService.upsertLineup(lineup)
 
         await matchmakingService.deleteLineupQueuesByChannelId(interaction.channelId)
-        const lineupQueue = await new LineupQueue({ lineup }).save()
+        const lineupQueue = await new LineupQueue({ lineup, ranked }).save()
 
         let reply = await interactionUtils.createReplyForLineup(interaction, lineup, lineupQueue)
         reply.embeds?.splice(0, 0, interactionUtils.createInformationEmbed(interaction.user, '✅ New mix lineup configured'))
